+++ src/main/java/com/ga/gaent/controller/NewsController.java
@@ -1,98 +1,28 @@
 package com.ga.gaent.controller;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.UnsupportedEncodingException;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.net.URLEncoder;
-import java.util.HashMap;
-import java.util.Map;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import com.ga.gaent.service.NewsService;
+import com.ga.gaent.util.TeamColor;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @RestController
 public class NewsController {
     
-    @Value("${navernews-clientid-key}") 
-    String clientId;
-    @Value("${navernews-clientsecret-key}") 
-    String clientSecret;
+    @Autowired NewsService newsService;
     
-    @GetMapping("/api/naverSearch") // Ajax 요청 처리
-    public String naverSearch(@RequestParam(name = "id") String search) {
+    @PostMapping("/api/naverSearch") // Ajax 요청 처리
+    public String naverSearch(
+            @RequestParam(name = "naverSearch") String search,
+            @RequestParam(name = "page", defaultValue = "1") int page
+            ) {
         
-        String text = search; // Ajax로 전달된 검색어 사용
-        try {
-            text = URLEncoder.encode(text, "UTF-8");
-        } catch (UnsupportedEncodingException e) {
-            throw new RuntimeException("검색어 인코딩 실패", e);
-        }
-
-        String apiURL = "https://openapi.naver.com/v1/search/news?query="
-                + text + "&display=" + 5 + "&start=" + 1 + "&sort=date"; // JSON 결과
-
-        Map<String, String> requestHeaders = new HashMap<>();
-        requestHeaders.put("X-Naver-Client-Id", clientId);
-        requestHeaders.put("X-Naver-Client-Secret", clientSecret);
-        String responseBody = get(apiURL, requestHeaders);
-
-        return responseBody;
+        log.debug(TeamColor.RED + "확인 : " + search + "& " + page + TeamColor.RESET);
+        
+        return newsService.naverSearch(search,page);
     }
-
-    private static String get(String apiUrl, Map<String, String> requestHeaders) {
-        HttpURLConnection con = connect(apiUrl);
-        try {
-            con.setRequestMethod("GET");
-            for (Map.Entry<String, String> header : requestHeaders.entrySet()) {
-                con.setRequestProperty(header.getKey(), header.getValue());
-            }
-
-            int responseCode = con.getResponseCode();
-            if (responseCode == HttpURLConnection.HTTP_OK) { // 정상 호출
-                return readBody(con.getInputStream());
-            } else { // 오류 발생
-                return readBody(con.getErrorStream());
-            }
-        } catch (IOException e) {
-            throw new RuntimeException("API 요청과 응답 실패", e);
-        } finally {
-            con.disconnect();
-        }
-    }
-
-    private static HttpURLConnection connect(String apiUrl) {
-        try {
-            URL url = new URL(apiUrl);
-            return (HttpURLConnection) url.openConnection();
-        } catch (MalformedURLException e) {
-            throw new RuntimeException("API URL이 잘못되었습니다. : " + apiUrl, e);
-        } catch (IOException e) {
-            throw new RuntimeException("연결이 실패했습니다. : " + apiUrl, e);
-        }
-    }
-
-    private static String readBody(InputStream body) {
-        InputStreamReader streamReader = new InputStreamReader(body);
-
-        try (BufferedReader lineReader = new BufferedReader(streamReader)) {
-            StringBuilder responseBody = new StringBuilder();
-
-            String line;
-            while ((line = lineReader.readLine()) != null) {
-                responseBody.append(line);
-            }
-
-            return responseBody.toString();
-        } catch (IOException e) {
-            throw new RuntimeException("API 응답을 읽는 데 실패했습니다.", e);
-        }
-    }
-    
-    
 }
