+++ src/main/webapp/WEB-INF/view/home/navernews-card.jsp
@@ -7,7 +7,7 @@ pageEncoding="UTF-8"%>
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <style>
       .news-card {
-        width: 34rem;
+        width: 36rem;
         height: 10rem;
       }
 
@@ -43,95 +43,130 @@ pageEncoding="UTF-8"%>
       .news-search-result hr {
         margin: 0.5rem 0;
       }
+      
+      .pagination {
+        margin-bottom: 0;
+      }
     </style>
   </head>
-  <body>
+<body>
     <div class="news-card">
-      <div class="card">
-        <div class="card-body">
-          <div class="d-flex align-items-center justify-content-between">
-            <div class="card-header">
-              <h4 class="card-title">네이버 뉴스</h4>
+        <div class="card">
+            <div class="card-body">
+                <div class="d-flex align-items-center justify-content-between">
+                    <div class="card-header">
+                        <h4 class="card-title">뉴스</h4>
+                    </div>
+                    <form id="searchForm">
+                        <div class="input-group">
+                            <input class="form-control from-control-sm" type="text" id="searchQuery" name="searchQuery" placeholder="검색어를 입력하세요"/>
+                            <button class="btn btn-outline-primary" id="naverNewsSearchBtn" type="submit">검색</button>
+                        </div>
+                    </form>
+                </div>
+                <hr />
+                <div id="searchResults" class="card-text news-search-result"></div>
+                <div class="d-flex justify-content-center">
+                    <nav aria-label="Page navigation">
+                        <ul class="pagination">
+                            <li id="prevNewsLi" class="page-item prev" >
+                                <button id="prevNewsBtn" class="page-link btn" >
+                                    <i class="tf-icon bx bx-chevron-left"></i>
+                                </button>
+                            </li>
+                            <li id="nextNewsLi" class="page-item next">
+                                <button id="nextNewsBtn" class="page-link btn btn-outline-primary">
+                                    <i class="tf-icon bx bx-chevron-right"></i>
+                                </button>
+                            </li>
+                        </ul>
+                    </nav>
+                </div>
             </div>
-            <form id="searchForm">
-              <div class="input-group">
-                <input
-                  class="form-control from-control-sm"
-                  type="text"
-                  id="searchQuery"
-                  name="searchQuery"
-                  placeholder="검색어를 입력하세요"
-                  value="엔터"
-                />
-                <button
-                  class="btn btn-outline-primary"
-                  id="naverNewsSearchBtn"
-                  type="submit"
-                >
-                  검색
-                </button>
-              </div>
-            </form>
-          </div>
-          <hr />
-          <div id="searchResults" class="card-text news-search-result"></div>
         </div>
-      </div>
     </div>
-    <script>
-      $(document).ready(function () {
+<script>
+    $(document).ready(function() {
+        let newsCurrentPage = 1;
+        let searchQuery = '엔터';  // 첫 검색어
+    
         // 페이지 로드 시 기본값인 '엔터'로 검색 결과 표시
-        performSearch('엔터');
-
+        performSearch(searchQuery, newsCurrentPage);
+    
         // 검색 폼 제출 시 AJAX를 통해 검색 요청
-        $('#searchForm').submit(function (event) {
-          event.preventDefault();
-          let searchQuery = $('#searchQuery').val();
-          performSearch(searchQuery);
+        $('#searchForm').submit(function(event) {
+            event.preventDefault();
+            searchQuery = $('#searchQuery').val();  // 전역 변수 업데이트
+            newsCurrentPage = 1;  // 검색 시 페이지를 1로 리셋
+            performSearch(searchQuery, newsCurrentPage);
+            updateNewPagingStates(newsCurrentPage);
         });
-      });
-
-      // 검색 API 호출 및 결과 표시
-      function performSearch(query) {
+    
+        $('#prevNewsBtn').click(function(event) {
+            event.preventDefault();
+            if (newsCurrentPage > 1) {  // 페이지가 1보다 클 때만 이전 페이지로 이동
+                newsCurrentPage--;
+                performSearch(searchQuery, newsCurrentPage);
+                updateNewPagingStates(newsCurrentPage);
+            }
+        });
+    
+        $('#nextNewsBtn').click(function(event) {
+            event.preventDefault();
+            newsCurrentPage++;
+            performSearch(searchQuery, newsCurrentPage);
+            updateNewPagingStates(newsCurrentPage);
+        });
+    });
+    
+    // 검색 API 호출 및 결과 표시
+    function performSearch(searchQuery, newsCurrentPage) {
         $.ajax({
-          type: 'GET',
-          url: '/gaent/api/naverSearch?id=' + query,
-          success: function (data) {
-            displayResults(data);
-          },
-          error: function () {
-            alert('검색에 실패했습니다.');
-          },
+            type: 'POST',
+            url: '/gaent/api/naverSearch',
+            data: {
+                naverSearch: searchQuery,
+                page: newsCurrentPage
+            },
+            success: function(data) {
+                displayResults(data);
+                updateNewPagingStates(newsCurrentPage);
+            },
+            error: function() {
+                alert('검색에 실패했습니다.');
+            },
         });
-      }
-
-      // 검색 결과를 화면에 출력
-      function displayResults(data) {
+    }
+    
+    // 검색 결과를 화면에 출력
+    function displayResults(data) {
         let results = JSON.parse(data);
         let resultDiv = $('#searchResults');
         resultDiv.empty();
         if (results.hasOwnProperty('items')) {
-          let items = results.items;
-          items.forEach(function (item) {
-            let title =
-              '<p class="news-title"><b><a href="' +
-              item.link +
-              '">' +
-              item.title +
-              '</a></b></p>';
-            // 원본 item.description에서 앞 40글자를 가져와서 "..."을 붙여 출력
-            let trimmedDescription =
-              item.description.substring(0, 56) +
-              (item.description.length > 40 ? '...' : '');
-            let description =
-              '<p class="news-desc">' + trimmedDescription + '</p>';
-            let hr = '<hr>';
-            resultDiv.append(title + description + hr);
-          });
+            let items = results.items;
+            items.forEach(function(item) {
+                let title = '<p class="news-title"><b><a href="' + item.link + '">' + item.title + '</a></b></p>';
+                // 원본 item.description에서 앞 45글자를 가져와서 "..."을 붙여 출력
+                let trimmedDescription = item.description.substring(0, 45) + (item.description.length > 45 ? '...' : '');
+                let description = '<p class="news-desc">' + trimmedDescription + '</p>';
+                let hr = '<hr>';
+                resultDiv.append(title + description + hr);
+            });
         } else {
-          resultDiv.append('<p>검색 결과가 없습니다.</p>');
+            resultDiv.append('<p>검색 결과가 없습니다.</p>');
         }
-      }
-    </script>
-  </body>
+    }
+    
+    // 버튼 상태 업데이트
+    function updateNewPagingStates(newsCurrentPage) {
+        if (newsCurrentPage <= 1) {
+            $('#prevNewsLi').addClass('disabled');
+        } else {
+            $('#prevNewsLi').removeClass('disabled');
+            $('#prevNewsBtn').addClass('btn-outline-primary');
+        }
+    }
+</script>
+</body>
 </html>
